fix: verify go mod
