+++ exchange-diary/infrastructure/configs/dev.yaml
